updated-4

+++ database/migrations/2025_04_10_144309_create_justificativa_faltas_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('justificativa_faltas', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('it_id_frequencia');
+            $table->unsignedBigInteger('it_id_frequencia');
             $table->foreign('it_id_frequencia')->references('id')->on('frequencias')->cascadeOnDelete();
             $table->String('vc_descricao');
             $table->timestamps();
